+++ vtm/src/org/oscim/utils/GeoPointUtils.java
@@ -15,12 +15,14 @@
 package org.oscim.utils;
 
 import org.oscim.core.GeoPoint;
+import org.oscim.core.Point;
 
 public final class GeoPointUtils {
 
     /**
-     * Find if the given point lies within this polygon.<br/>
-     * See http://www.ecse.rpi.edu/Homepages/wrf/Research/Short_Notes/pnpoly.html
+     * Find if the given point lies within this polygon.
+     * <p>
+     * http://www.ecse.rpi.edu/Homepages/wrf/Research/Short_Notes/pnpoly.html
      *
      * @return true if this polygon contains the given point, false otherwise.
      */
@@ -37,12 +39,38 @@ public final class GeoPointUtils {
     }
 
     /**
+     * Returns the distance between the given segment and point.
+     * <p>
+     * libGDX (Apache 2.0)
+     */
+    public static double distanceSegmentPoint(double startX, double startY, double endX, double endY, double pointX, double pointY) {
+        Point nearest = nearestSegmentPoint(startX, startY, endX, endY, pointX, pointY);
+        return Math.hypot(nearest.x - pointX, nearest.y - pointY);
+    }
+
+    /**
      * Find if this way is closed.
      *
      * @return true if this way is closed, false otherwise.
      */
     public static boolean isClosedWay(GeoPoint[] geoPoints) {
         return geoPoints[0].distance(geoPoints[geoPoints.length - 1]) < 0.000000001;
+    }
+
+    /**
+     * Returns a point on the segment nearest to the specified point.
+     * <p>
+     * libGDX (Apache 2.0)
+     */
+    public static Point nearestSegmentPoint(double startX, double startY, double endX, double endY, double pointX, double pointY) {
+        double xDiff = endX - startX;
+        double yDiff = endY - startY;
+        double length2 = xDiff * xDiff + yDiff * yDiff;
+        if (length2 == 0) return new Point(startX, startY);
+        double t = ((pointX - startX) * (endX - startX) + (pointY - startY) * (endY - startY)) / length2;
+        if (t < 0) return new Point(startX, startY);
+        if (t > 1) return new Point(endX, endY);
+        return new Point(startX + t * (endX - startX), startY + t * (endY - startY));
     }
 
     private GeoPointUtils() {